blando/app/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.4
django-rest-framework==0.1.0
django-solo==1.1.5
djangorestframework==3.12.2
Pillow==8.0.1
pycodestyle==2.6.0
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
